+++ ui/action.h
@@ -1,0 +1,9 @@
+#ifndef UI_ACTION_H
+#define UI_ACTION_H
+
+#include <gtk/gtk.h>
+
+bool action_init(GtkApplication *app, GtkWidget* bar);
+void action_update(void);
+
+#endif