+++ ui/appbar.h
@@ -4,6 +4,7 @@
 #include <gtk/gtk.h>
 
 bool appbar_init(GtkApplication *app, GtkWidget *win);
+void appbar_update(void);
 extern guint datetime_id;
 
 #endif